third_party/blink/web_tests/inspector-protocol/css/css-add-rule-expected.txt

The test verifies functionality of protocol method CSS.addRule.
==== Initial style sheet text ====
#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}



Running test: testAddRuleToStyleSheetBeginning
==== Style sheet text ====
#test { content: 'EDITED'; }#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    content: 'EDITED'; @[0:8-0:26]
    content: "EDITED"; @[undefined-undefined]
}
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testAddRuleToStyleSheetEnding
==== Style sheet text ====
#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}
#test { content: 'EDITED'; }

Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
*#test* {    regular
    content: 'EDITED'; @[20:8-20:26]
    content: "EDITED"; @[undefined-undefined]
}
Dumping inherited rules: 

Running test: testAddRuleToStyleSheetCenter
==== Style sheet text ====
#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}
#test { content: 'EDITED'; }
@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
*#test* {    regular
    content: 'EDITED'; @[11:8-11:26]
    content: "EDITED"; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testAddRuleToRuleEnding
==== Style sheet text ====
#test {
    box-sizing: border-box;
}#test{
    content: 'EDITED';
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    content: 'EDITED'; @[3:4-3:22]
    content: "EDITED"; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[8:4-8:19]
    font-family: "Arial"; @[9:4-9:25]
    color: blue; @[10:4-10:16]
    display: flex; @[11:4-11:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[16:8-16:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testAddRuleWithFormatting
==== Style sheet text ====
#test {
    box-sizing: border-box;
}

#test{
    content: 'EDITED';
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    content: 'EDITED'; @[5:4-5:22]
    content: "EDITED"; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[10:4-10:19]
    font-family: "Arial"; @[11:4-11:25]
    color: blue; @[12:4-12:16]
    display: flex; @[13:4-13:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[18:8-18:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testAddRuleToMediaRuleBeginning
==== Style sheet text ====
#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {#test { content: 'EDITED'; }
    #test {
        font-size: 200%;
    }

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        content: 'EDITED'; @[12:33-12:51]
        content: "EDITED"; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testAddRuleToMediaRuleCenter
==== Style sheet text ====
#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }
#test { content: 'EDITED'; }
    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
        content: 'EDITED'; @[16:8-16:26]
        content: "EDITED"; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testAddRuleToMediaRuleEnd
==== Style sheet text ====
#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
#test { content: 'EDITED'; }}


Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
@media (min-width: 1px)
    *#test* {    regular
        content: 'EDITED'; @[19:8-19:26]
        content: "EDITED"; @[undefined-undefined]
    }
Dumping inherited rules: 

Running test: testInvalidRule
Expected protocol error: SyntaxError Rule text is not valid.

Running test: testInvalidRule2
Expected protocol error: SyntaxError Rule text is not valid.

Running test: testInvalidRule3
Expected protocol error: SyntaxError Rule text is not valid.

Running test: testAddingRuleInsideSelector
Expected protocol error: NotFoundError Cannot insert rule inside rule selector.

Running test: testAddingRuleBeforeRuleBody
Expected protocol error: NotFoundError Cannot insert rule inside rule selector.

Running test: testAddingRuleInsideMedia1
Expected protocol error: NotFoundError Cannot insert rule inside rule selector.

Running test: testAddingRuleInsideMedia2
Expected protocol error: NotFoundError Cannot insert rule inside rule selector.

Running test: testAddingRuleBeforeMediaBody
Expected protocol error: NotFoundError Cannot insert rule inside rule selector.

Running test: testAddingRuleInsideStyleRule
Expected protocol error: NotFoundError Cannot insert rule in non-media rule.
